requirements.txt
# Ambient Learning Agent System — Python dependencies
uagents>=0.13.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
google-genai>=1.0.0
pydantic>=2.5.0
websockets>=12.0
pytest>=7.4.0
